Create README.md
What Ido:
Adaptation of documentation to parity and non-sexist language.
Advice in easy language for documentation.
Documentary reports and spinning of cases for advocacy.
Office without papers. 
Implementation of the system to comply with Data Protection.
Creation of teleworking systems. 
Audit and implementation of systems for Knowledge Management. ISO and safety.
Ontology.